refactor: make TakeABreak a module
